# backend/tests/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "Dispute": {
        "description": "A case. status: open, auto_resolved (the policy resolved it), escalated (waiting for a specialist), in_progress (a specialist took it) or closed.",
        "properties": {
          "case_id": {
            "format": "uuid",
            "title": "Case Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_id": {
            "title": "Customer Id",
            "type": "string"
          },
          "evidence": {
            "additionalProperties": true,
            "description": "The transaction as it was, plus resolution (auto_resolved) or handoff (escalated).",
            "title": "Evidence",
            "type": "object"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null for an escalation that could not be tied to a transaction.",
            "title": "Transaction Id"
          }
        },
        "required": [
          "case_id",
          "customer_id",
          "reason_code",
          "summary",
          "status",
          "evidence",
          "created_at"
        ],
        "title": "Dispute",
        "type": "object"
      },
      "DisputeCreate": {
        "properties": {
          "reason_code": {
            "description": "e.g. unrecognized_charge, duplicate, failed, other",
            "minLength": 1,
            "title": "Reason Code",
            "type": "string"
          },
          "summary": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "transaction_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          }
        },
        "required": [
          "reason_code",
          "summary"
        ],
        "title": "DisputeCreate",
        "type": "object"
      },
      "DisputeDetail": {
        "properties": {
          "case_id": {
            "format": "uuid",
            "title": "Case Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_id": {
            "title": "Customer Id",
            "type": "string"
          },
          "events": {
            "description": "Append-only timeline: created, then resolved or escalated.",
            "items": {
              "$ref": "#/components/schemas/DisputeEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "evidence": {
            "additionalProperties": true,
            "description": "The transaction as it was, plus resolution (auto_resolved) or handoff (escalated).",
            "title": "Evidence",
            "type": "object"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null for an escalation that could not be tied to a transaction.",
            "title": "Transaction Id"
          }
        },
        "required": [
          "case_id",
          "customer_id",
          "reason_code",
          "summary",
          "status",
          "evidence",
          "created_at",
          "events"
        ],
        "title": "DisputeDetail",
        "type": "object"
      },
      "DisputeEvent": {
        "properties": {
          "event": {
            "title": "Event",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          }
        },
        "required": [
          "event",
          "payload",
          "ts"
        ],
        "title": "DisputeEvent",
        "type": "object"
      },
      "EscalateRequest": {
        "properties": {
          "handoff": {
            "additionalProperties": true,
            "description": "Structured handoff payload for a human agent (request, verified facts, actions taken, evidence, open questions)",
            "title": "Handoff",
            "type": "object"
          }
        },
        "required": [
          "handoff"
        ],
        "title": "EscalateRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResolveRequest": {
        "properties": {
          "resolution": {
            "description": "no_charge_confirmed for Declined, reversal_confirmed for Reversed",
            "pattern": "^(no_charge_confirmed|reversal_confirmed)$",
            "title": "Resolution",
            "type": "string"
          }
        },
        "required": [
          "resolution"
        ],
        "title": "ResolveRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\nMock banking service (tool layer) for the Factored AI & Data Hackathon 2026\nprototype. Sandbox limitations, stated explicitly:\n\n- Login is a test session: `POST /session` checks `customer_id` +\n  `document_number` against `gold.customers` and issues a short-lived JWT\n  (HS256). No real identity provider is behind it.\n- Admin login (`POST /admin/session`) checks credentials against `ADMIN_USERS`\n  (bcrypt hashes, in-memory rate limit, single-process sandbox limitation).\n  Admins work an inbox of escalated cases with audited transitions\n  (claim -> in_progress -> closed); they never set `auto_resolved`.\n- `GET /meta/demo-scenarios` is a public demo aid: it surfaces only the fields\n  the test login itself asks for (customer_id, document_number, name), never\n  fraud columns.\n- Reads come from the read-only `gold.*` tables published by the dbt pipeline;\n  operational tables (`app.*`) are built by the versioned SQL migrations in\n  `app/migrations`, applied at startup under an advisory lock.\n- `/v1` is what the web app uses (user name and password login, products,\n  paginated history, summary). The routes at the root are the agent's.\n- No money is ever moved: disputes only explain, document and escalate.\n- Fraud ground truth (`is_fraud`/`fraud_score`) is not part of `gold.transactions`\n  and is never exposed here.\n",
    "title": "mock-banking-service",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/disputes": {
      "get": {
        "operationId": "list_disputes_admin_disputes_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "customer_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Disputes Admin Disputes Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Disputes",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/disputes/{case_id}": {
      "get": {
        "operationId": "get_dispute_admin_disputes__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Dispute Admin Disputes  Case Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dispute",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/disputes": {
      "get": {
        "description": "The customer's own cases, newest first.",
        "operationId": "list_disputes_disputes_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(open|auto_resolved|escalated|in_progress)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Dispute"
                  },
                  "title": "Response List Disputes Disputes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Disputes",
        "tags": [
          "disputes"
        ]
      },
      "post": {
        "description": "Open the case. One per (customer, transaction): reporting the same transaction again returns that case.",
        "operationId": "create_dispute_disputes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispute"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
    "/disputes/{case_id}": {
      "get": {
        "operationId": "get_dispute_disputes__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
    "/disputes/{case_id}/escalate": {
      "post": {
        "operationId": "escalate_dispute_disputes__case_id__escalate_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispute"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Escalate Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
    "/disputes/{case_id}/resolve": {
      "post": {
        "description": "System (agent) resolution after verifying the case: 'open' -> 'auto_resolved'.\nOnly the safe path ever calls this; humans close cases through /admin transitions.",
        "operationId": "resolve_dispute_disputes__case_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispute"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
...
    "/me/disputes": {
      "get": {
        "operationId": "list_my_disputes_me_disputes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List My Disputes Me Disputes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List My Disputes",
        "tags": [
          "customers"
        ]
      }
    },
...
    "/v1/disputes": {
      "get": {
        "description": "The customer's own cases, newest first.",
        "operationId": "list_disputes_v1_disputes_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(open|auto_resolved|escalated|in_progress)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Dispute"
                  },
                  "title": "Response List Disputes V1 Disputes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Disputes",
        "tags": [
          "disputes"
        ]
      },
      "post": {
        "description": "Open the case. One per (customer, transaction): reporting the same transaction again returns that case.",
        "operationId": "create_dispute_v1_disputes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispute"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
    "/v1/disputes/{case_id}": {
      "get": {
        "operationId": "get_dispute_v1_disputes__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
    "/v1/disputes/{case_id}/escalate": {
      "post": {
        "operationId": "escalate_dispute_v1_disputes__case_id__escalate_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispute"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Escalate Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
    "/v1/disputes/{case_id}/resolve": {
      "post": {
        "description": "System (agent) resolution after verifying the case: 'open' -> 'auto_resolved'.\nOnly the safe path ever calls this; humans close cases through /admin transitions.",
        "operationId": "resolve_dispute_v1_disputes__case_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dispute"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve Dispute",
        "tags": [
          "disputes"
        ]
      }
    },
...
  }
}